% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_surface_markers}
\alias{example_surface_markers}
\title{example_surface_markers}
\format{
A tibble with 1000 rows (cells) and 28 columns (markers)
}
\source{
Internal data from Marie Burns, at the German Rheumatism Research Center
}
\usage{
data("example_surface_markers")
}
\description{
This is the surface markers of a whole blood mass cytometry dataset.
}
