% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_umap}
\alias{example_umap}
\title{example_umap}
\format{
A tibble with 1000 rows (cells) and 2 columns (umap dimensions)
}
\source{
Internal data from Marie Burns, at the German Rheumatism Research Center
}
\usage{
data("example_umap")
}
\description{
This is the umap run on the surface markers of the aforementioend whole
blood mass cytometry dataset.
}
