% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_sleepwalk_functions.R
\name{GridSamplePoints}
\alias{GridSamplePoints}
\title{Sample representative points on a grid}
\usage{
GridSamplePoints(mat, xdim = 10, ydim = 10)
}
\arguments{
\item{mat}{A data matrix of points in a 2-D embedding, with points as
rows}

\item{xdim}{Grid width}

\item{ydim}{Grid height}
}
\value{
Numeric vector
}
\description{
Returns indices of representative points within a 2-D
embedding, sampled uniformly on a \code{xdim}-by-\code{ydim} grid. Points closest to
the center of each non-empty tile are selected.
}
\examples{
## Load example embedding & convert to matrix
data("example_umap") # `umap`
umap <- as.matrix(umap)

## Sample representative points
pts <- GridSamplePoints(umap)

## Show these points in a plot
plot(umap, pch = 20, cex = .5, col = "gray", asp = 1,
     xlab = "Component 1", ylab = "Component 2")
points(umap[pts, 1], umap[pts, 2],
       pch = 10, cex = 2, col = "darkred")

## Load the original data
data("example_surface_markers") # `surface`

## Compute local neighborhoods in high dimension
nn <- MakeNnMatrix(surface, k = 100, ranks = TRUE, rann = TRUE)

## Create composite PNG of plots of neighborhoods for the sample points
neighb_plots <- lapply(pts, function(pt) 
  PlotNeighborhood(embedding = umap, point_index = pt, nn = nn) +
    theme(legend.position = "none"))
figure <- cowplot::plot_grid(plotlist = neighb_plots)
ggplot2::ggsave("./neighbors.png", figure,
                width = 30, height = 30, dpi = 300)
}
\seealso{
\link{PlotNeighborhood} to plot local neighborhoods of
}
