% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_sleepwalk_functions.R
\name{KfnSleepwalk}
\alias{KfnSleepwalk}
\title{K-farthest neighbors Sleepwalk Direct}
\usage{
KfnSleepwalk(
  mat1,
  mat2,
  embedding,
  k = 100,
  output_file = NULL,
  point_size = 1.5,
  plot_names = c("KFN embedding space", "KFN high-dim space")
)
}
\arguments{
\item{mat1}{A data matrix with data points as rows and features as columns.
the matrix must already be filtered by the markers you care about.}

\item{mat2}{A second data matrix in the format as above}

\item{embedding}{The 2-D embedding of the data matrix, in matrix format with
data points as rows and two columns.}

\item{k}{The number of nearest neighbors to be visualized}

\item{output_file}{The file to save your sleepwalk html page to}

\item{point_size}{How big you want the point on the plots to be}

\item{plot_names}{What you want the comparison plots to be named}
}
\description{
Takes a data matrix and a 2-D embedding as input. It produces
a 'KFN matrix' and places that along with the aforementioned inputs into the
sleepwalk function. Unlike the default KFN sleepwalk function, this one makes
two KFN matrices, like UMAP space vs high-D space for comparisons across the
same embedding.
}
