% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_sleepwalk_functions.R
\name{PlotNeighborhood}
\alias{PlotNeighborhood}
\title{Plot point neighborhood in an embedding}
\usage{
PlotNeighborhood(embedding, point_index, nn, point_size = 1)
}
\arguments{
\item{embedding}{A data matrix of points in a 2-D embedding, with points as
rows}

\item{point_index}{Row index of a single point within \code{embedding}}

\item{nn}{Neighborhood matrix computed using \code{KnnSleepwalk::MakeNnMatrix}
with \code{ranks=TRUE}}

\item{point_size}{size of points in the embedding plot}

\item{preserve_aspect_ratio}{whether to fix a 1-to-1 aspect ratio between the
latent-space components}
}
\value{
A \code{ggplot} object
}
\description{
The neighborhood of a single point is shown in an embedding.
This is useful for visualizing the preservation of local structure and
comparing that between embeddings. To that end, the neighborhood matrix
should be computed for the original data.
}
\examples{
## Load example embedding & convert to matrix
data("example_umap") # `umap`
umap <- as.matrix(umap)

## Sample representative points
pts <- GridSamplePoints(umap)

## Show these points in a plot
plot(umap, pch = 20, cex = .5, col = "gray", asp = 1,
     xlab = "Component 1", ylab = "Component 2")
points(umap[pts, 1], umap[pts, 2],
       pch = 10, cex = 2, col = "darkred")

## Load the original data
data("example_surface_markers") # `surface`

## Compute local neighborhoods in high dimension
nn <- MakeNnMatrix(surface, k = 100, ranks = TRUE, rann = TRUE)

## Create composite PNG of plots of neighborhoods for the sample points
neighb_plots <- lapply(pts, function(pt) 
  PlotNeighborhood(embedding = umap, point_index = pt, nn = nn) +
    theme(legend.position = "none"))
figure <- cowplot::plot_grid(plotlist = neighb_plots)
ggplot2::ggsave("./neighbors.png", figure,
                width = 30, height = 30, dpi = 300)

}
\seealso{
\link{GridSamplePoints} to sample representative points for
use in the \code{point_index}
}
