% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_sleepwalk_functions.R
\name{SleepwalkSP}
\alias{SleepwalkSP}
\title{K-NN Sleepwalk view of Local Structure Preservation}
\usage{
SleepwalkSP(
  embeddings,
  nn,
  plot_names = paste0("Emb", seq_along(embeddings)),
  output_file = NULL,
  point_size = 1.5,
  metric = "euclidean"
)
}
\arguments{
\item{embeddings}{Single embedding in matrix format with data points
as rows and two columns, or a list of multiple embeddings}

\item{nn}{Nearest-neighbor matrix of original data generated using
\code{KnnSleepwalk::MakeNnMatrix} with \code{ranks=TRUE}}

\item{plot_names}{What you want the comparison plots to be named. This vector
of strings needs to match the number of plots (eg. 2 if we're dealing with
2 plots here).}

\item{output_file}{The file to save your sleepwalk html page to}

\item{point_size}{How big you want the point on the plots to be}
}
\description{
Assesses local structure preservation (SP) in a 2-D embedding by
projecting K-nearest-neighbor ranks from the original data onto it. If
multiple embeddings are specified, it compares them side-by-side with respect
to point-wise Local SP.
}
\examples{
## Load example data & UMAP embedding
data("example_umap") # `umap`
data("example_surface_markers") # `surface`

## Generate alternative PCA embedding
pca <- prcomp(surface)$x[, 1:2]

## Compute neighbor matrix
nn <- MakeNnMatrix(surface, k = 100, ranks = TRUE, rann = TRUE)

## Compare structure preservation by PCA & UMAP
SleepwalkSP(
    embeddings = list(pca, umap),
    plot_names = c("PCA", "UMAP"),
    nn = nn,
    point_size = 3
)
}
\seealso{
\link{MakeNnMatrix} to pre-compute the neighbor matrix
}
