% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_sleepwalk_functions.R
\name{BiaxialSleepwalk}
\alias{BiaxialSleepwalk}
\title{Biaxial Sleepwalk}
\usage{
BiaxialSleepwalk(
  root_biax,
  biax_list,
  k = 100,
  output_file = NULL,
  point_size = 1.5,
  plot_names = c()
)
}
\arguments{
\item{root_biax}{The biaxial plot from which the KNN are calculated}

\item{biax_list}{A list that starts with root_biax and continues with the
remaining biaxials that you care about. These will be visualized.}

\item{k}{The number of nearest neighbors}

\item{output_file}{The file that the output gets saved to. If set to NULL,
then no save takes place}

\item{point_size}{The size of the points on the plots.}

\item{plot_names}{These will place names on top of the plots. This vector of
strings must equal the total number of plots being visualized.}
}
\description{
A generalization of KNN sleepwalk, whereby the K-nearest
neighbors of the first biaxial plot are calculated, and the corresponding
cells are highlighted on any of the other biaxials that are on the map
}
\examples{
data("example_umap")
data("example_surface_markers")
BiaxialSleepwalk(
  root_biax = surface[,c("CD4", "CD8")],
  biax_list = list(
    surface[,c("CD4", "CD8")],
    surface[,c("IgA", "IgM")],
    surface[,c("IgD", "HLA-DR")]
  )
)
}
