% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_sleepwalk_functions.R
\name{KnnSleepwalk}
\alias{KnnSleepwalk}
\title{K-nearest neighbors Sleepwalk}
\usage{
KnnSleepwalk(
  mat,
  embedding,
  k = 100,
  output_file = NULL,
  point_size = 1.5,
  dimr_names = c("t-SNE", "UMAP")
)
}
\arguments{
\item{mat}{A data matrix with data points as rows and features as columns.
the matrix must already be filtered by the markers you care about.}

\item{embedding}{The 2-D embedding of the data matrix, in matrix format with
data points as rows and two columns.}

\item{k}{The number of nearest neighbors to be visualized}
}
\description{
Takes a data matrix and a 2-D embedding as input. It produces
a 'KNN matrix' and places that along with the aforementioned inputs into the
sleepwalk function.
}
