% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_sleepwalk_functions.R
\name{MakeNnMatrix}
\alias{MakeNnMatrix}
\title{Make nearest neighbor matrix}
\usage{
MakeNnMatrix(mat, kfn = FALSE, k = 100, metric = "euclidean")
}
\arguments{
\item{mat}{A data matrix, presumably from a single-cell dataset}

\item{kfn}{Whether you want to look at the K-nearest or K-farthest neighbors}

\item{k}{The number of nearest neighbors}

\item{metric}{The distance metric to be used. Available options: euclidean,
manhattan, and cosine}
}
\value{
The aforementioned nearest neighbor matrix
}
\description{
A helper function for KnnSleepwalk which takes a matrix as
input, computes the distance matrix, and then changes the values such that
the output of the KnnSleepwalk will show the K-nearest or fartherst neighbors
as black, with the rest of the cells being a lighter color. This is what we
call the nearest neighbor matrix.
}
\keyword{internal}
